pyproject.toml
[tool.poetry]
name = "ipumspy"
version = "0.1.0"
description = "A collection of tools for working with IPUMS data"
authors = ["Kevin H. Wilson <kevin_wilson@brown.edu>"]
license = "MIT"
packages = [
    { include = "ipumspy", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.6.1"
pandas = "^1.0.4"
click = "^7.0.0"
pyarrow = "^3.0.0"
requests = "^2.25.1"
importlib-metadata = {version = "^4.0.1", python = "<3.8"}

[tool.poetry.dev-dependencies]
black = "^19.10b0"
pylint = "^2.7.4"
isort = "^5.8.0"
mypy = "^0.812"
pytest = "^6.2.3"
python-dotenv = "^0.17.0"

[tool.poetry.scripts]
ipums = "ipumspy.cli:cli"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88

[tool.pylint.basic]
good-names = "i,j,k,ex,Run,_,df,pc"

[tool.pylint.messages_control]
disable = "C0330, C0326, R0912, R0913, R0914, R0915"

[tool.pylint.format]
max-line-length = "88"

[tool.pytest.ini_options]
markers = [ "slow: Run with --runslow" ]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
